ubication permission fix

+++ CarShop/CarShop/Models/Car.cs
@@ -14,7 +14,7 @@ namespace CarShop.Models
         public int Year { get; set; }
         public string PhotoUrl { get; set; }
 
-        public float? lat { get; set; }
-        public float? Lon { get; set; }
+        public Double? lat { get; set; }
+        public Double? Lon { get; set; }
     }
 }
